+++ public/index.html
@@ -7,6 +7,7 @@
     <title>react webpack</title>
   </head>
   <body>
+    <img src="./webreactjs.png" alt="reactor setup github" width="500" height="300">
     <div class="root" id="root"></div>
   </body>
 </html>
