+++ src/App.jsx
@@ -1,21 +1,14 @@
-import React, {useCallback, useState} from 'react';
+import React from 'react';
+import Counter from './components/Counter.jsx';
+
+// implementig lazy loading
+
+const Counter = React.lazy(()=>import('./components/Counter'));
 
 const App = () => {
-  const [count, setCount] = useState(0);
-
-  const increment = useCallback(() => {
-    setCount(count + 1);
-  }, [count]);
-
-  const decrement = useCallback(() => {
-    setCount(count - 1);
-  }, [count])
   return (
     <div className="app">
-      <h1>welcome to webpack react!</h1>
-      <h2>{count}</h2>
-      <button onClick={decrement}>-</button>
-      <button onClick={increment}>+</button>
+      <Counter/>
     </div>
   );
 };
