+++ src/components/Counter.jsx
@@ -1,0 +1,22 @@
+import React, { useState, useCallback } from "react";
+
+const Counter = () => {
+  const [count, setCount] = useState(0);
+
+  const increment = useCallback(() => {
+    setCount(count + 1);
+  }, [count]);
+
+  const decrement = useCallback(() => {
+    setCount(count - 1);
+  }, [count]);
+  return (
+    <div className="counter">
+      <h2>{count}</h2>
+      <button onClick={decrement}>-</button>
+      <button onClick={increment}>+</button>
+    </div>
+  );
+};
+
+export default Counter;
