+++ src/styles/index.css
@@ -1,3 +1,3 @@
 body{
-  background-color: aqua;
+  background-color: rgb(255, 255, 255);
 }
